requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
idna==2.8
inflection==0.3.1
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.1
more-itertools==5.0.0
ndg-httpsclient==0.5.1
numpy==1.16.4
pandas==0.24.2
pyasn1==0.4.6
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
Quandl==3.4.8
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
urllib3==1.25.3
